refactor: enforce kebab-case filenames
Rename the 21 remaining camelCase/PascalCase files and message directories
to kebab-case per AGENTS.md, and update every importer. Exported symbol
names are unchanged, and the `next-intl` namespaces still live inside the
message JSON, so `useTranslations('articleIndex')` and
`useTranslations('cookiePreferences')` keep working untouched.

Re-applied from #1011 (chore/remove-dead-code) against current main.
Dropped the `apps/site/tsconfig.json` hunk (main already generalised the
Storybook include to `.storybook/**/*.ts{,x}`), the deleted
`(contact)/actions.ts` importer hunks (main replaced contact with a Sanity
form), and the `ui/drawer.tsx` / `lib/types/components.ts` deletions plus
the `vaul` removal (still imported by the legacy marketing header, which
this branch deliberately keeps).
