+++ apps/sanity/schema-types/article-shared-fields.ts
