+++ apps/site/src/app/(unauthenticated)/[locale]/(marketing)/components/header/header.stories.tsx
@@ -5,7 +5,7 @@ import Header from './header';
 import {
   storybookControls,
   storybookArgs,
-} from '../../../../../../../.storybook/utils/storybookControls';
+} from '../../../../../../../.storybook/utils/storybook-controls';
 
 const meta: Meta<typeof Header> = {
   title: 'UI/Header',
