+++ apps/site/src/components/common/news/article-share-copy-link.tsx
@@ -7,7 +7,7 @@ import {
   TooltipContent,
   TooltipTrigger,
 } from '@/components/ui/tooltip';
-import useCurrentUrl from '@/lib/hooks/useCurrentUrl';
+import useCurrentUrl from '@/lib/hooks/use-current-url';
 import logger from '@/lib/logger';
 import { useState } from 'react';
 import { BiLinkAlt } from 'react-icons/bi';
