+++ apps/site/src/i18n/message-files.ts
@@ -12,13 +12,13 @@ export const messageFiles = [
   'homepage',
   'about',
   'research',
-  'articleIndex',
+  'article-index',
   'investors',
   'governance',
   'accessibility',
   'privacy',
   'terms',
   'careers',
-  'cookiePreferences',
+  'cookie-preferences',
   'forms',
 ];
