+++ apps/site/src/test/test-utils.tsx
@@ -1,6 +1,6 @@
 // Copyright © Todd Agriscience, Inc. All rights reserved.
 
-import { ThemeProvider } from '@/context/theme/ThemeContext';
+import { ThemeProvider } from '@/context/theme/theme-context';
 import { act, fireEvent, render, RenderOptions } from '@testing-library/react';
 import { NextIntlClientProvider } from 'next-intl';
 import React from 'react';
